requirements.txt
django-ckeditor
pillow
pytils
segno
faker
django-random-queryset